+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * [EACQAPW-5332] CardListPresenter Tests
 * [EACQAPW-5347] CardPaymentProcess Tests
 * [EACQAPW-5362] CellImageLoader Tests
+* [EACQAPW-5371] TDSCertsManager Tests
 
 ### Fixed